backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
dnspython==2.7.0
bcrypt==4.2.1
PyJWT==2.10.1
requests==2.32.3
python-dotenv==1.0.1
